restructure frontend files a bit
